feat(core):extract update selections and collision logic to hook
